Added cast to int to spinner value
git-svn-id: https://svn.code.sf.net/p/sweethome3d/code/trunk/SweetHome3D@717 03e323e1-6c6c-46a5-9f71-c4ce5d8c6204

+++ test/com/eteks/sweethome3d/junit/HomeCameraTest.java
@@ -269,7 +269,7 @@ public class HomeCameraTest extends ComponentTestFixture {
     int oldSkyColor = home.getSkyColor();
     int oldLightColor = home.getLightColor();
     float oldWallsAlpha = home.getWallsAlpha();
-    assertEquals("Wrong field of view", Math.round(Math.toDegrees(oldCameraFieldOfView)), 
+    assertEquals("Wrong field of view", (int)Math.round(Math.toDegrees(oldCameraFieldOfView)), 
         observerFieldOfViewSpinner.getValue());
     assertEquals("Wrong height", (float)Math.round(oldCameraHeight * 100) / 100, 
         observerHeightSpinner.getValue());
